requirements.txt
Flask>=2.2
pyTelegramBotAPI>=4.11.0
requests>=2.28
Pillow>=9.0
google-genai>=1.0.0
gunicorn>=20.1
python-dotenv>=1.0